requirements.txt
openai
langchain
fastapi
uvicorn
python-dotenv
langchain-core
langchain-openai
faiss-cpu
tiktoken
python-dotenv
requests
beautifulsoup4
langchain-community